Flip logic from dpkg output
Hi there!  I was having some issues with the nag window persisting after running your role (with the correct variable set), and the fix on my end was modifying this script to look for an exit status of 0 instead of 1.

The dpkg command looks for proxmoxlib.js, and then if that's successful, the sed command should run.  With '$? -eq 1', it would only run if the file was not found, which seems like it would be counterproductive.

Am I missing something here by chance?